+++ streams/src/test/java/org/apache/kafka/streams/integration/utils/IntegrationTestUtils.java
@@ -49,7 +49,7 @@ import java.util.Properties;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
-import static org.apache.kafka.streams.processor.internals.ConsumerUtil.poll;
+import static org.apache.kafka.streams.processor.internals.ConsumerUtils.poll;
 
 /**
  * Utility functions to make integration testing more convenient.
