+++ streams/src/test/java/org/apache/kafka/streams/perf/SimpleBenchmark.java
@@ -62,7 +62,7 @@ import java.util.Random;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-import static org.apache.kafka.streams.processor.internals.ConsumerUtil.poll;
+import static org.apache.kafka.streams.processor.internals.ConsumerUtils.poll;
 
 /**
  * Class that provides support for a series of benchmarks. It is usually driven by
